refactor: change name of imported component
